pyproject.toml
[tool.poetry]
name = "cal-magic"
version = "0.1.0"
description = ""
authors = ["Sujit Kamthe <sujitk@sahaj.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-index-readers-google = "^0.4.0"
openai = "^1.42.0"
protobuf = "3.20.0"
langchain-text-splitters = "^0.2.2"
langchain = "^0.2.14"
langchain-community = "^0.2.12"
langchain-core = "^0.2.34"
langchain-chroma = "^0.1.3"
sentence-transformers = "^3.0.1"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
